pyproject.toml
[tool.poetry]
name = "django-practice"
version = "0.1.0"
description = ""
authors = ["pacho-h <pacho.h90@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.1"
Django = "^5.0.3"
django-environ = "^0.11.2"
djangorestframework = "^3.15.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.3.2"
pytest = "^8.1.1"
pytest-watcher = "^0.4.1"
pynvim = "^0.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "ANN",  # flake8-annotations
    "D",  # pydocstyle
    "EM",  # flake8-errmsg
    "TRY003",  # tryceratops
]
